+++ CMakeLists.txt
@@ -150,13 +150,13 @@ add_library(Vex STATIC
     "src/Vex/CommandQueueType.h"
     "src/Vex/CommandContext.h"
     "src/Vex/CommandContext.cpp"
-    "src/Vex/ShaderKey.h"
+    "src/Vex/Shaders/ShaderKey.h"
     "src/Vex/Bindings.h"
     "src/Vex/Bindings.cpp"
     "src/Vex/PipelineStateCache.h"
     "src/Vex/PipelineStateCache.cpp"
-    "src/Vex/Shader.h"
-    "src/Vex/Shader.cpp"
+    "src/Vex/Shaders/Shader.h"
+    "src/Vex/Shaders/Shader.cpp"
     "src/Vex/Hash.h"
     "src/Vex/Containers/FreeList.h"
     "src/Vex/Containers/FreeList.cpp"
@@ -164,21 +164,25 @@ add_library(Vex STATIC
     "src/Vex/Resource.h"
     "src/Vex/Containers/ResourceCleanup.h"
     "src/Vex/Containers/ResourceCleanup.cpp"
-    "src/Vex/ShaderCompiler.h"
-    "src/Vex/ShaderCompiler.cpp"
+    "src/Vex/Shaders/ShaderCompiler.h"
+    "src/Vex/Shaders/ShaderCompiler.cpp"
     "src/Vex/Formattable.h"
-    "src/Vex/ShaderResourceContext.h"
-    "src/Vex/ShaderResourceContext.cpp"
-    "src/Vex/ShaderGen.h"
+    "src/Vex/Shaders/ShaderResourceContext.h"
+    "src/Vex/Shaders/ShaderResourceContext.cpp"
+    "src/Vex/Shaders/ShaderGen.h"
     "src/Vex/TextureSampler.h"
     "src/Vex/GraphicsPipeline.h"
     "src/Vex/DrawHelpers.h"
-    "src/Vex/ShaderCompilerSettings.h"
+    "src/Vex/Shaders/ShaderCompilerSettings.h"
     "src/Vex/RenderExtension.h"
     "src/Vex/RenderExtension.cpp"
     "src/Vex/ResourceBindingUtils.cpp"
     "src/Vex/MaybeUninitialized.h"
     "src/Vex/NonNullPtr.h"
+    "src/Vex/Shaders/DXCImpl.h"
+    "src/Vex/Shaders/DXCImpl.cpp"
+    "src/Vex/Shaders/ShaderEnvironment.h"
+    "src/Vex/Shaders/CompilerInterface.h"
     "src/Vex/RayTracing.h"
     "src/Vex/RayTracing.cpp"
     "src/Vex/ByteUtils.h"
