+++ src/DX12/RHI/DX12RHI.h
@@ -43,8 +43,8 @@ public:
 
     virtual RHIAllocator CreateAllocator() override;
 
-    virtual void ModifyShaderCompilerEnvironment(std::vector<const wchar_t*>& args,
-                                                 std::vector<ShaderDefine>& defines) override;
+    virtual void ModifyShaderCompilerEnvironment(ShaderCompilerBackend compilerBackend,
+                                                 ShaderEnvironment& shaderEnv) override;
 
     virtual u32 AcquireNextFrame(RHISwapChain& swapChain, u32 currentFrameIndex) override;
     virtual void SubmitAndPresent(std::span<RHICommandList*> commandLists,
