+++ src/RHI/RHIPipelineState.h
@@ -4,8 +4,8 @@
 #include <Vex/Hash.h>
 #include <Vex/RHIFwd.h>
 #include <Vex/RayTracing.h>
-#include <Vex/Shader.h>
-#include <Vex/ShaderKey.h>
+#include <Vex/Shaders/Shader.h>
+#include <Vex/Shaders/ShaderKey.h>
 #include <Vex/Types.h>
 
 namespace vex
