+++ src/Vex/Shaders/ShaderCompiler.cpp
@@ -10,9 +10,12 @@
 #include <Vex/PhysicalDevice.h>
 #include <Vex/Platform/Platform.h>
 #include <Vex/RHIImpl/RHI.h>
-#include <Vex/Shader.h>
-#include <Vex/ShaderGen.h>
-#include <Vex/ShaderResourceContext.h>
+#include <Vex/Shaders/CompilerInterface.h>
+#include <Vex/Shaders/DXCImpl.h>
+#include <Vex/Shaders/Shader.h>
+#include <Vex/Shaders/ShaderEnvironment.h>
+#include <Vex/Shaders/ShaderGen.h>
+#include <Vex/Shaders/ShaderResourceContext.h>
 #include <Vex/TextureSampler.h>
 
 namespace vex
@@ -20,57 +23,6 @@ namespace vex
 
 namespace ShaderCompiler_Internal
 {
-
-static std::vector<DxcDefine> ConvertDefinesToDxcDefine(const std::vector<ShaderDefine>& defines)
-{
-    std::vector<DxcDefine> dxcDefines;
-    dxcDefines.reserve(defines.size());
-    for (auto& d : defines)
-    {
-        dxcDefines.emplace_back(DxcDefine{ .Name = d.name.c_str(), .Value = d.value.c_str() });
-    }
-    return dxcDefines;
-}
-
-static std::wstring GetTargetFromShaderType(ShaderType type)
-{
-    FeatureChecker* featureChecker = GPhysicalDevice->featureChecker.get();
-
-    std::wstring highestSupportedShaderModel =
-        StringToWString(std::string(magic_enum::enum_name(featureChecker->GetShaderModel())));
-
-    using enum ShaderType;
-
-    // First char changes depending on shader type.
-    switch (type)
-    {
-    case VertexShader:
-        highestSupportedShaderModel[0] = L'v';
-        break;
-    case PixelShader:
-        highestSupportedShaderModel[0] = L'p';
-        break;
-    case ComputeShader:
-        highestSupportedShaderModel[0] = L'c';
-        break;
-    // Raytracing shaders use "lib_*" target profile
-    case ShaderType::RayGenerationShader:
-    case ShaderType::RayMissShader:
-    case ShaderType::RayClosestHitShader:
-    case ShaderType::RayAnyHitShader:
-    case ShaderType::RayIntersectionShader:
-    case ShaderType::RayCallableShader:
-        return L"lib" + highestSupportedShaderModel.substr(2);
-    default:
-        VEX_LOG(Fatal, "Unsupported shader type for the Vex ShaderCompiler.");
-        break;
-    }
-
-    // Second character is always 's' for non-RT shaders.
-    highestSupportedShaderModel[1] = L's';
-
-    return highestSupportedShaderModel;
-}
 
 struct ShaderParser
 {
@@ -303,27 +255,7 @@ private:
 
 } // namespace ShaderCompiler_Internal
 
-CompilerUtil::CompilerUtil()
-{
-    // Create compiler and utils.
-    HRESULT hr1 = DxcCreateInstance(CLSID_DxcCompiler, IID_PPV_ARGS(&compiler));
-    HRESULT hr2 = DxcCreateInstance(CLSID_DxcUtils, IID_PPV_ARGS(&utils));
-
-    if (FAILED(hr1) || FAILED(hr2))
-    {
-        VEX_LOG(Fatal, "Failed to create DxcCompiler and/or DxcUtils...");
-    }
-
-    // Create default include handler.
-    HRESULT hr3 = utils->CreateDefaultIncludeHandler(&defaultIncludeHandler);
-
-    if (FAILED(hr3))
-    {
-        VEX_LOG(Fatal, "Failed to create default include handler...");
-    }
-}
-
-CompilerUtil::~CompilerUtil() = default;
+ShaderCompiler::ShaderCompiler() = default;
 
 ShaderCompiler::ShaderCompiler(RHI* rhi, const ShaderCompilerSettings& compilerSettings)
     : rhi(rhi)
@@ -333,47 +265,24 @@ ShaderCompiler::ShaderCompiler(RHI* rhi, const ShaderCompilerSettings& compilerS
     // Force disable shader debugging in shipping.
     this->compilerSettings.enableShaderDebugging = false;
 #endif
+
+    switch (compilerSettings.compilerBackend)
+    {
+    case ShaderCompilerBackend::DXC:
+    {
+        compilerImpl = MakeUnique<DXCCompilerImpl>();
+        break;
+    }
+    }
 }
 
 ShaderCompiler::~ShaderCompiler() = default;
-
-ComPtr<IDxcResult> ShaderCompiler::GetPreprocessedShader(const Shader& shader,
-                                                         const ComPtr<IDxcBlobEncoding>& shaderBlobUTF8) const
-{
-    std::vector<LPCWSTR> args = { L"-P" }; // -P gives us the preprocessor output of the shader.
-    FillInAdditionalIncludeDirectories(args);
-
-    DxcBuffer buffer = { .Ptr = shaderBlobUTF8->GetBufferPointer(),
-                         .Size = shaderBlobUTF8->GetBufferSize(),
-                         .Encoding = CP_UTF8 };
-
-    ComPtr<IDxcResult> result;
-    if (HRESULT hr = GetCompilerUtil().compiler->Compile(&buffer,
-                                                         args.data(),
-                                                         static_cast<u32>(args.size()),
-                                                         nullptr,
-                                                         IID_PPV_ARGS(&result));
-        FAILED(hr))
-    {
-        return nullptr;
-    }
-
-    return result;
-}
-
-void ShaderCompiler::FillInAdditionalIncludeDirectories(std::vector<LPCWSTR>& args) const
-{
-    for (auto& additionalIncludeFolder : additionalIncludeDirectories)
-    {
-        args.insert(args.end(), { L"-I", StringToWString(additionalIncludeFolder.string()).c_str() });
-    }
-}
 
 std::expected<std::string, std::string> ShaderCompiler::ProcessShaderCodeGen(
     const std::string& shaderFileStr, const ShaderResourceContext& resourceContext)
 {
     using ShaderCompiler_Internal::ShaderParser;
-    // We insert the local constants on the vulkan side by replacing the first declaration of VEX_LOCAL_CONSTANTS,
+    // We insert the local constants on the Vulkan side by replacing the first declaration of VEX_LOCAL_CONSTANTS,
     // however this macro should only appear once and has some other constraints. This parser validates the usage of
     // VEX_LOCAL_CONSTANTS in the user shader code.
     auto res = ShaderParser::ParseShader(shaderFileStr);
@@ -475,36 +384,18 @@ std::expected<std::string, std::string> ShaderCompiler::ProcessShaderCodeGen(
     return ShaderParser::ReplaceVexShaderBlock(shaderFileStr, shaderBlockInfo, codeGen);
 }
 
-CompilerUtil& ShaderCompiler::GetCompilerUtil()
+ShaderEnvironment ShaderCompiler::CreateShaderEnvironment()
 {
-    static thread_local CompilerUtil GCompilerUtil;
-    return GCompilerUtil;
-}
+    ShaderEnvironment env;
+    env.defines.emplace_back(L"VEX_DEBUG", std::to_wstring(VEX_DEBUG));
+    env.defines.emplace_back(L"VEX_DEVELOPMENT", std::to_wstring(VEX_DEVELOPMENT));
+    env.defines.emplace_back(L"VEX_SHIPPING", std::to_wstring(VEX_SHIPPING));
+    env.defines.emplace_back(L"VEX_RAYTRACING",
+                             GPhysicalDevice->featureChecker->IsFeatureSupported(Feature::RayTracing) ? L"1" : L"0");
 
-std::optional<std::size_t> ShaderCompiler::GetShaderHash(const Shader& shader) const
-{
-    ComPtr<IDxcBlobEncoding> shaderBlobUTF8;
-    u32 codePage = CP_UTF8;
-    if (HRESULT hr = GetCompilerUtil().utils->LoadFile(shader.key.path.wstring().c_str(), &codePage, &shaderBlobUTF8);
-        FAILED(hr))
-    {
-        VEX_LOG(Error, "Unable to get shader hash, failed to load shader from filepath: {}", shader.key.path.string());
-        return std::nullopt;
-    }
+    rhi->ModifyShaderCompilerEnvironment(compilerSettings.compilerBackend, env);
 
-    if (ComPtr<IDxcResult> preprocessingResult = GetPreprocessedShader(shader, shaderBlobUTF8))
-    {
-        ComPtr<IDxcBlob> preprocessedShader;
-        if (HRESULT hr = preprocessingResult->GetOutput(DXC_OUT_HLSL, IID_PPV_ARGS(&preprocessedShader), nullptr);
-            SUCCEEDED(hr))
-        {
-            std::string_view finalSourceView{ static_cast<const char*>(preprocessedShader->GetBufferPointer()),
-                                              preprocessedShader->GetBufferSize() };
-            return std::hash<std::string_view>{}(finalSourceView);
-        }
-    }
-
-    return std::nullopt;
+    return env;
 }
 
 std::expected<void, std::string> ShaderCompiler::CompileShader(Shader& shader,
@@ -518,7 +409,7 @@ std::expected<void, std::string> ShaderCompiler::CompileShader(Shader& shader,
     // Generate the hash if this is the first time we've compiled this shader.
     if (shader.version == 0)
     {
-        std::optional<std::size_t> newHash = GetShaderHash(shader);
+        std::optional<std::size_t> newHash = compilerImpl->GetShaderHash(shader, additionalIncludeDirectories);
         if (!newHash)
         {
             return std::unexpected("Failed to generate shader hash.");
@@ -534,6 +425,7 @@ std::expected<void, std::string> ShaderCompiler::CompileShader(Shader& shader,
     }
     std::string shaderFileStr = buffer.str();
 
+    // Shader code-gen, should work in any compilation backend, if not I'll move it into the impls later on.
     {
         auto res = ProcessShaderCodeGen(shaderFileStr, resourceContext);
         if (!res.has_value())
@@ -547,149 +439,17 @@ std::expected<void, std::string> ShaderCompiler::CompileShader(Shader& shader,
     VEX_LOG(Verbose, "Shader {}\nFile dump:\n{}", shader.key, shaderFileStr);
 #endif
 
-    ComPtr<IDxcBlobEncoding> shaderBlob;
-    if (HRESULT hr = GetCompilerUtil().utils->CreateBlobFromPinned(shaderFileStr.c_str(),
-                                                                   shaderFileStr.size(),
-                                                                   CP_UTF8,
-                                                                   &shaderBlob);
-        FAILED(hr))
+    ShaderEnvironment env = CreateShaderEnvironment();
+
+    auto res = compilerImpl->CompileShader(shaderFileStr, env, additionalIncludeDirectories, compilerSettings, shader);
+    if (!res.has_value())
     {
-        return std::unexpected("Failed to create blob from shader.");
+        return std::unexpected(res.error());
     }
-
-    DxcBuffer shaderSource = {};
-    shaderSource.Ptr = shaderBlob->GetBufferPointer();
-    shaderSource.Size = shaderBlob->GetBufferSize();
-    shaderSource.Encoding = DXC_CP_ACP; // Assume BOM says UTF8 or UTF16 or this is ANSI text.
-
-    std::vector<ShaderDefine> shaderDefines = shader.key.defines;
-    shaderDefines.emplace_back(L"VEX_DEBUG", std::to_wstring(VEX_DEBUG));
-    shaderDefines.emplace_back(L"VEX_DEVELOPMENT", std::to_wstring(VEX_DEVELOPMENT));
-    shaderDefines.emplace_back(L"VEX_SHIPPING", std::to_wstring(VEX_SHIPPING));
-    shaderDefines.emplace_back(L"VEX_RAYTRACING",
-                               GPhysicalDevice->featureChecker->IsFeatureSupported(Feature::RayTracing) ? L"1" : L"0");
-
-    std::vector<LPCWSTR> args;
-    rhi->ModifyShaderCompilerEnvironment(args, shaderDefines);
-
-    if (compilerSettings.enableShaderDebugging)
-    {
-        args.insert(args.end(),
-                    {
-                        DXC_ARG_DEBUG,
-                        DXC_ARG_WARNINGS_ARE_ERRORS,
-                        DXC_ARG_DEBUG_NAME_FOR_SOURCE,
-                        L"-Qembed_debug",
-                    });
-    }
-
-    if (compilerSettings.enableHLSL202xFeatures)
-    {
-        args.insert(args.end(), HLSL202xFlags.begin(), HLSL202xFlags.end());
-    }
-
-    FillInAdditionalIncludeDirectories(args);
-
-    std::vector<DxcDefine> defines = ShaderCompiler_Internal::ConvertDefinesToDxcDefine(shaderDefines);
-
-    std::wstring entryPoint;
-    // RT Shaders are compiled differently to other shader types, the entry point should be null.
-    if (!IsRayTracingShader(shader.key.type))
-    {
-        entryPoint = StringToWString(shader.key.entryPoint);
-    }
-
-    ComPtr<IDxcCompilerArgs> compilerArgs;
-    if (HRESULT hr = GetCompilerUtil().utils->BuildArguments(
-            shader.key.path.wstring().c_str(),
-            entryPoint.empty() ? nullptr : entryPoint.c_str(),
-            ShaderCompiler_Internal::GetTargetFromShaderType(shader.key.type).c_str(),
-            args.data(),
-            static_cast<u32>(args.size()),
-            defines.data(),
-            static_cast<u32>(defines.size()),
-            &compilerArgs);
-        FAILED(hr))
-    {
-        return std::unexpected(
-            std::format("Failed to build shader compilation arguments from file: {}", shader.key.path.string()));
-    }
-
-    ComPtr<IDxcResult> shaderCompilationResults;
-    HRESULT compilationHR = GetCompilerUtil().compiler->Compile(&shaderSource,
-                                                                compilerArgs->GetArguments(),
-                                                                compilerArgs->GetCount(),
-                                                                GetCompilerUtil().defaultIncludeHandler.operator->(),
-                                                                IID_PPV_ARGS(&shaderCompilationResults));
-
-    ComPtr<IDxcBlobUtf8> errors = nullptr;
-    if (HRESULT hrError = shaderCompilationResults->GetOutput(DXC_OUT_ERRORS, IID_PPV_ARGS(&errors), nullptr);
-        FAILED(hrError))
-    {
-        return std::unexpected("Failed to compile shader due to unknown reasons, the DXC compilation error "
-                               "was unable to be obtained.");
-    }
-    if (errors != nullptr && errors->GetStringLength() != 0)
-    {
-        return std::unexpected(std::format("{}", errors->GetStringPointer()));
-    }
-
-    ComPtr<IDxcBlob> shaderBytecode;
-    HRESULT objectResult = shaderCompilationResults->GetOutput(DXC_OUT_OBJECT, IID_PPV_ARGS(&shaderBytecode), nullptr);
-    if (FAILED(objectResult))
-    {
-        return std::unexpected("Failed to obtain the shader blob after compilation.");
-    }
+    std::vector<u8>& shaderBytecode = res.value();
 
     // Store shader bytecode blob inside the Shader.
-    shader.blob.resize(shaderBytecode->GetBufferSize());
-    std::memcpy(shader.blob.data(), shaderBytecode->GetBufferPointer(), shader.blob.size() * sizeof(u8));
-
-    // TODO(https://trello.com/c/9D3KOgHr): Reflection blob, to be implemented later on.
-    {
-        // ComPtr<IDxcBlob> reflectionBlob;
-        // HRESULT reflectionResult =
-        //     shaderCompilationResults->GetOutput(DXC_OUT_REFLECTION, IID_PPV_ARGS(&reflectionBlob),
-        //     nullptr);
-        // if (FAILED(reflectionResult))
-        // {
-        //     return std::unexpected("Unable to get the shader reflection blob after compilation.");
-        // }
-        //
-        // DxcBuffer reflectionBuffer;
-        // reflectionBuffer.Encoding = DXC_CP_ACP;
-        // reflectionBuffer.Ptr = reflectionBlob->GetBufferPointer();
-        // reflectionBuffer.Size = reflectionBlob->GetBufferSize();
-
-        // ComPtr<ID3D12ShaderReflection> reflectionData;
-        // chk << compilerUtil.utils->CreateReflection(&reflectionBuffer, IID_PPV_ARGS(&reflectionData));
-
-        // D3D12_SHADER_DESC reflectionDesc;
-        // reflectionData->GetDesc(&reflectionDesc);
-        // for (uint32 i = 0; i < reflectionDesc.BoundResources; ++i)
-        //{
-        //     D3D12_SHADER_INPUT_BIND_DESC bindDesc;
-        //     reflectionData->GetResourceBindingDesc(i, &bindDesc);
-
-        //    switch (bindDesc.Type)
-        //    {
-        //    case D3D_SIT_CBUFFER:
-        //    case D3D_SIT_TEXTURE:
-        //    case D3D_SIT_STRUCTURED:
-        //    case D3D_SIT_BYTEADDRESS:
-        //    case D3D_SIT_UAV_RWTYPED:
-        //    case D3D_SIT_UAV_RWSTRUCTURED:
-        //    case D3D_SIT_UAV_RWBYTEADDRESS:
-        //        cbvSrvUavReflectionMap[std::string(bindDesc.Name)] = bindDesc.BindPoint;
-        //        break;
-        //    case D3D_SIT_SAMPLER:
-        //        samplerReflectionMap[std::string(bindDesc.Name)] = bindDesc.BindPoint;
-        //        break;
-        //    default:
-        //        break;
-        //    }
-        //}
-    }
+    shader.blob = std::move(shaderBytecode);
 
     shader.version++;
     shader.isDirty = false;
@@ -740,7 +500,7 @@ std::pair<bool, std::size_t> ShaderCompiler::IsShaderStale(const Shader& shader)
         return { false, shader.hash };
     }
 
-    std::optional<std::size_t> newHash = GetShaderHash(shader);
+    std::optional<std::size_t> newHash = compilerImpl->GetShaderHash(shader, additionalIncludeDirectories);
     if (!newHash)
     {
         return { false, shader.hash };
