+++ src/Vex/Shaders/ShaderEnvironment.h
@@ -1,0 +1,17 @@
+#pragma once
+
+#include <string>
+#include <vector>
+
+namespace vex
+{
+
+struct ShaderDefine;
+
+struct ShaderEnvironment
+{
+    std::vector<std::wstring> args;
+    std::vector<ShaderDefine> defines;
+};
+
+} // namespace vex
