+++ src/Vex/Shaders/ShaderResourceContext.cpp
